Update column
git-svn-id: file:///home/svn/acars2006@16338 1e6d2d28-e1fb-0310-b8ab-81908d8c6b6a

+++ sql/create_sysinfo.sql
@@ -8,7 +8,7 @@ CREATE TABLE SYSINFO (
  SLI                        BOOLEAN NOT NULL DEFAULT FALSE,
  LOCALE                     VARCHAR(12) NOT NULL,
  TZ                         VARCHAR(64) NOT NULL,
- MEMORY                     SMALLINT UNSIGNED NOT NULL,
+ MEMORY                     INTEGER UNSIGNED NOT NULL,
  CPU                        VARCHAR(48) NOT NULL,
  CPU_SPEED                  SMALLINT UNSIGNED NOT NULL,
  CPU_SOCK                   SMALLINT UNSIGNED NOT NULL,
